+++ metatensor-torch/include/metatensor/torch/misc.hpp
@@ -26,26 +26,44 @@ namespace details {
 /// Load a previously saved `TensorMap` from the given path.
 METATENSOR_TORCH_EXPORT TorchTensorMap load(const std::string& path);
 
-/// Load previously saved `Labels` from the given path.
-METATENSOR_TORCH_EXPORT TorchLabels load_labels(const std::string& path);
-
 /// Load a previously saved `TensorMap` from the given in-memory buffer
 /// (represented as a `torch::Tensor` of bytes)
 METATENSOR_TORCH_EXPORT TorchTensorMap load_buffer(torch::Tensor buffer);
+
+/// Save the given `TensorMap` to a file at `path`
+METATENSOR_TORCH_EXPORT void save(const std::string& path, TorchTensorMap tensor);
+
+/// Save the given `TensorMap` to an in-memory buffer (represented as a
+/// `torch::Tensor` of bytes)
+METATENSOR_TORCH_EXPORT torch::Tensor save_buffer(TorchTensorMap tensor);
+
+/******************************************************************************/
+
+/// Load a previously saved `TensorBlock` from the given path.
+METATENSOR_TORCH_EXPORT TorchTensorBlock load_block(const std::string& path);
+
+/// Load a previously saved `TensorBlock` from the given in-memory buffer
+/// (represented as a `torch::Tensor` of bytes)
+METATENSOR_TORCH_EXPORT TorchTensorBlock load_block_buffer(torch::Tensor buffer);
+
+/// Save the given `TensorBlock` to a file at `path`
+METATENSOR_TORCH_EXPORT void save(const std::string& path, TorchTensorBlock block);
+
+/// Save the given `TensorBlock` to an in-memory buffer (represented as a
+/// `torch::Tensor` of bytes)
+METATENSOR_TORCH_EXPORT torch::Tensor save_buffer(TorchTensorBlock block);
+
+/******************************************************************************/
+
+/// Load previously saved `Labels` from the given path.
+METATENSOR_TORCH_EXPORT TorchLabels load_labels(const std::string& path);
 
 /// Load previously saved `Labels` from the given in-memory buffer
 /// (represented as a `torch::Tensor` of bytes)
 METATENSOR_TORCH_EXPORT TorchLabels load_labels_buffer(torch::Tensor buffer);
 
-/// Save the given `TensorMap` to a file at `path`
-METATENSOR_TORCH_EXPORT void save(const std::string& path, TorchTensorMap tensor);
-
 /// Save the given `Labels` to a file at `path`
 METATENSOR_TORCH_EXPORT void save(const std::string& path, TorchLabels labels);
-
-/// Save the given `TensorMap` to an in-memory buffer (represented as a
-/// `torch::Tensor` of bytes)
-METATENSOR_TORCH_EXPORT torch::Tensor save_buffer(TorchTensorMap tensor);
 
 /// Save the given `Labels` to an in-memory buffer (represented as a
 /// `torch::Tensor` of bytes)